requirements.txt
argparse==1.2.1
github3.py==0.9.3
requests==2.5.1
uritemplate.py==0.3.0
wsgiref==0.1.2
